requirements.txt
absl-py==0.15.0
altair==4.2.2
astor==0.8.1
astunparse==1.6.3
attrs==25.4.0
base58==2.1.1
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cycler==0.12.1
entrypoints==0.4
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.43.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.11
importlib_metadata==8.7.0
imutils==0.5.4
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.17.3
jsonschema-specifications==2025.9.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.4.1
mdurl==0.1.2
numpy==1.19.5
oauthlib==3.3.1
onnx==1.10.1
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==25.0
pandas==1.4.4
Pillow==8.3.2
protobuf==3.20.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pyparsing==3.2.5
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.30.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
smmap==5.0.2
streamlit==1.22.0
tenacity==8.5.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tf2onnx==1.9.3
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.4
wrapt==1.12.1
zipp==3.23.0
